+++ src/components-v2/VaultSearchControls/VaultSearchBar.tsx
@@ -1,4 +1,4 @@
-import React, { KeyboardEvent } from 'react';
+import React from 'react';
 import { IconButton, InputAdornment, makeStyles, TextField } from '@material-ui/core';
 import SearchIcon from '@material-ui/icons/Search';
 
@@ -14,24 +14,17 @@ const useStyles = makeStyles({
 interface Props {
 	search?: string;
 	onChange: (search: string) => void;
-	onSearch: () => void;
 }
 
-const VaultSearchBar = ({ search = '', onChange, onSearch }: Props): JSX.Element => {
+const VaultSearchBar = ({ search = '', onChange }: Props): JSX.Element => {
 	const classes = useStyles();
-
-	const handleKeyPress = (event: KeyboardEvent<HTMLInputElement>) => {
-		if (event.key === 'Enter') {
-			onSearch();
-		}
-	};
 
 	return (
 		<TextField
 			InputProps={{
 				endAdornment: (
 					<InputAdornment position="end">
-						<IconButton size="small" aria-label="search" onClick={onSearch} className={classes.icon}>
+						<IconButton size="small" aria-label="search" className={classes.icon}>
 							<SearchIcon />
 						</IconButton>
 					</InputAdornment>
@@ -43,7 +36,6 @@ const VaultSearchBar = ({ search = '', onChange, onSearch }: Props): JSX.Element
 			className={classes.root}
 			value={search}
 			placeholder="Search by vault name, token, rewards..."
-			onKeyPress={handleKeyPress}
 			onChange={(e) => onChange(e.target.value)}
 		/>
 	);