+++ src/components-v2/vault-detail/TopContent.tsx
@@ -2,9 +2,16 @@ import React from 'react';
 import { Grid, makeStyles } from '@material-ui/core';
 import { Breadcrumb } from './Breadcrumb';
 import { Description } from './description/Description';
-import { VaultDTO } from '@badger-dao/sdk';
+import { VaultDTO, VaultState } from '@badger-dao/sdk';
+import VaultDeprecationWarning from '../VaultDeprecationWarning';
 
 const useStyles = makeStyles((theme) => ({
+	root: {
+		marginBottom: theme.spacing(4),
+		[theme.breakpoints.down('sm')]: {
+			marginBottom: theme.spacing(2),
+		},
+	},
 	content: {
 		margin: 'auto',
 		[theme.breakpoints.up('md')]: {
@@ -13,10 +20,6 @@ const useStyles = makeStyles((theme) => ({
 	},
 	descriptionSection: {
 		justifyContent: 'space-between',
-		marginBottom: theme.spacing(4),
-		[theme.breakpoints.down('sm')]: {
-			marginBottom: theme.spacing(2),
-		},
 	},
 	breadcrumbContainer: {
 		marginBottom: theme.spacing(1),
@@ -33,14 +36,23 @@ interface Props {
 export const TopContent = ({ vault }: Props): JSX.Element => {
 	const classes = useStyles();
 
+	console.log(vault.vaultToken);
+
 	return (
-		<Grid container>
-			<Grid container className={classes.breadcrumbContainer}>
-				<Breadcrumb vault={vault} />
+		<Grid container className={classes.root}>
+			<Grid item container xs>
+				<Grid container className={classes.breadcrumbContainer}>
+					<Breadcrumb vault={vault} />
+				</Grid>
+				<Grid container className={classes.descriptionSection}>
+					<Description vault={vault} />
+				</Grid>
 			</Grid>
-			<Grid container className={classes.descriptionSection}>
-				<Description vault={vault} />
-			</Grid>
+			{vault.state === VaultState.Discontinued && (
+				<Grid item container xs>
+					<VaultDeprecationWarning vault={vault} />
+				</Grid>
+			)}
 		</Grid>
 	);
 };